fix: address round 3 review findings
Architecture:
- Extract ChatActionManager to src/utils/chatAction.ts (72 lines)
- Extract probeMediaFile, generateThumbnail, extractStreamInfo to
  src/services/mediaProbe.ts (111 lines)
- Extract buildSingleCaption, buildGroupCaption to src/utils/caption.ts
  (67 lines)
- message.ts reduced from 560 to 360 lines; processMediaRequest split
  into fetchMediaInfo + downloadMedia helpers (each <50 lines)

Security/Correctness:
- Remove double withRetry in getYtDlpInfo (was 5×3=15 spawns, now 5)
- Add unhandledRejection + uncaughtException handlers in index.ts
- assertSafePath: reject filePath === baseDir (rel='')
- Lowercase SUPPORTED_DOMAINS at parse time (case mismatch fix)
- Guard gallery-dl tuple destructuring (2-element fallback)
- Validate gallery-dl output is non-empty array
- extractFilePath: trimEnd on captured paths (CRLF edge case)
- Bot permission check: add 'creator' status alongside 'administrator'
- parseInt on ffprobe bit_rate/sample_rate: NaN-safe parseBitrate helper
- generateThumbnail: delete temp file after readFile (no disk leak)
- cleanupFiles: only add .thumb.jpg for video items (no spurious unlinks)
- ChatActionManager maxDelay reduced to 3s (within Telegram's 5s window)

Infrastructure:
- Remove deprecated version: '3.8' from docker-compose.yml
- Create .env.example with placeholder values
- Logger: consistent process.stderr.write for error/warn,
  process.stdout.write for info/debug; precomputed level rank

Downloader:
- Use yt-dlp --print after_move:filepath instead of fragile stdout
  parsing (eliminates extractFilePath regex entirely)
- Use --quiet --no-warnings for clean stdout
