+++ .env.example
@@ -1,0 +1,7 @@
+# Required
+BOT_TOKEN=your_telegram_bot_token_here
+
+# Optional cookie files (for authenticated downloads)
+# COOKIES_FILE_YOUTUBE=/cookies/youtube.txt
+# COOKIES_FILE_INSTAGRAM=/cookies/instagram.txt
+# COOKIES_FILE_TWITTER=/cookies/twitter.txt
