+++ docker-compose.yml
@@ -1,5 +1,3 @@
-version: '3.8'
-
 services:
   tgmr:
     container_name: tgmr
